Correción getOcurrence

+++ Practicas/LAB01/Practica 01.c
@@ -95,7 +95,7 @@ void GENFUN_vCapsOff (uint8 *pu8Src, uint8 u8SizeOfList)
 	}
 }
 
-uint32 GENFUN_u8GetOccurence (uint8 *pu8Src, uint8 u8Target, uint8 u8SizeOfList) 
+uint8 GENFUN_u8GetOccurence (uint8 *pu8Src, uint8 u8Target, uint8 u8SizeOfList) 
 {
 	int Contrador=0;
